Fixed various bugs. Improvements

+++ Controller/LabelController.php
@@ -4,19 +4,24 @@ namespace DpdLabel\Controller;
 
 
 use DpdLabel\DpdLabel;
-use DpdLabel\Form\ApiConfigurationForm;
 use DpdLabel\Form\LabelGenerationForm;
 use DpdLabel\Model\DpdlabelLabelsQuery;
 use DpdLabel\Service\LabelService;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\ResponseHeaderBag;
+use Symfony\Component\Routing\Annotation\Route;
 use Thelia\Controller\Admin\BaseAdminController;
+use Thelia\Core\HttpFoundation\Request;
 use Thelia\Core\Security\AccessManager;
 use Thelia\Core\Security\Resource\AdminResources;
+use Thelia\Core\Translation\Translator;
+use Thelia\Exception\TheliaProcessException;
 use Thelia\Model\OrderQuery;
 use Thelia\Tools\URL;
-use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Route("/admin/module/DpdLabel", name="dpdlabel")
@@ -26,115 +31,108 @@ class LabelController extends BaseAdminController
     /**
      * @Route("/labels", name="_labels", methods="GET")
      */
-    public function showAction(RequestStack $requestStack)
+    public function showAction(Request $request): Response
     {
-        $err =  $requestStack->getCurrentRequest()->get("err");
         return $this->render('dpdlabel-labels', [
-            "err" => $err
+            "err" => $request->get("err")
         ]);
     }
 
     /**
      * @Route("/saveLabel", name="_save_label", methods="GET")
-     * @throws \Propel\Runtime\Exception\PropelException
      */
-    public function saveAction(RequestStack $requestStack, LabelService $labelService)
+    public function saveAction(Request $request, LabelService $labelService, Translator $translator): Response
     {
-        $request = $requestStack->getCurrentRequest();
         $orderId = $request->get("orderId");
-
-        $labelDir = DpdLabel::DPD_LABEL_DIR;
 
         $fileSystem = new Filesystem();
 
-        if (!$fileSystem->exists($labelDir)) {
-            $fileSystem->mkdir($labelDir, 0777);
+        if (!$fileSystem->exists(DpdLabel::DPD_LABEL_DIR)) {
+            $fileSystem->mkdir(DpdLabel::DPD_LABEL_DIR);
         }
-        $order = OrderQuery::create()->filterById($orderId)->findOne();
 
-        $labelName = $labelDir . DS . $order->getRef();
-        $labelName = $labelService->setLabelNameExtension($labelName);
-
-        if (!$label = dpdlabelLabelsQuery::create()->filterByOrderId($order->getId())->findOne()) {
-            $baseForm = $this->createForm(LabelGenerationForm::getName());
-            try {
-                $form = $this->validateForm($baseForm);
-                $data = $form->getData();
-            } catch (\Exception $e) {
-                return $this->generateRedirect(URL::getInstance()->absoluteUrl("admin/module/DpdLabel/labels", [
-                    "err" => $e->getMessage()
-                ]));
+        try {
+            if (null === $order = OrderQuery::create()->filterById($orderId)->findOne()) {
+                return new Response(
+                    $translator->trans(
+                        "Order ID %id not found",
+                        [ '%id' => $orderId ],
+                        DpdLabel::DOMAIN_NAME
+                    ),
+                    Response::HTTP_NOT_FOUND
+                );
             }
 
-            $err = $labelService->createLabel($order, $labelName, $data['weight']);
+            $labelName = DpdLabel::DPD_LABEL_DIR . $order->getRef();
+            $labelName = $labelService->setLabelNameExtension($labelName);
 
-            if (is_string($err)) {
-                return $this->generateRedirect(URL::getInstance()->absoluteUrl("admin/module/DpdLabel/labels", [
-                    "err" => $err
-                ]));
+            if (null !== DpdlabelLabelsQuery::create()->filterByOrderId($order->getId())->findOne()) {
+                return $this->downloadAction(base64_encode($labelName), $translator);
             }
+
+            $data = $this->validateForm($this->createForm(LabelGenerationForm::getName()))->getData();
+
+            $labelService->createLabel($order, $labelName, (float) $data['weight']);
 
             $params = ['file' => base64_encode($labelName)];
 
-            return $this->generateRedirect(URL::getInstance()->absoluteUrl('admin/module/DpdLabel/labels', $params));
-
+            return $this->generateRedirect(URL::getInstance()?->absoluteUrl('admin/module/DpdLabel/labels', $params));
+        } catch (\Exception $ex) {
+            return $this->generateRedirect(URL::getInstance()?->absoluteUrl("admin/module/DpdLabel/labels", [
+                "err" => $ex->getMessage()
+            ]));
         }
-
-        return $this->downloadAction(base64_encode($labelName));
     }
 
     /**
-     * @return \Symfony\Component\HttpFoundation\Response
-     * @throws \Propel\Runtime\Exception\PropelException
-     * @throws \SoapFault
      * @Route("/generateLabel", name="_generate_label", methods="POST")
      */
-    public function generateLabelAction(RequestStack $requestStack, LabelService $labelService)
+    public function generateLabelAction(Request $request, LabelService $labelService): Response
     {
-        $request = $requestStack->getCurrentRequest();
         $orderId = $request->get("orderId");
-        $retour = $request->get("retour");
+        $retour = (bool) $request->get("retour");
+        $returnUrl = $request->get("return_url");
+        $error = null;
 
-        $order = OrderQuery::create()->filterById($orderId)->findOne();
-
-        $labelDir = DpdLabel::DPD_LABEL_DIR;
-        $labelName = $labelDir . DS . $order->getRef();
-        $labelName = $labelService->setLabelNameExtension($labelName);
-
-        $fileSystem = new Filesystem();
-
-        if (!$fileSystem->exists($labelDir)) {
-            $fileSystem->mkdir($labelDir, 0777);
-        }
-
-        $baseForm = $this->createForm(LabelGenerationForm::getName());
         try {
-            $form = $this->validateForm($baseForm);
-            $data = $form->getData();
+            if (null === $order = OrderQuery::create()->filterById($orderId)->findOne()) {
+                throw new TheliaProcessException("Cannot find order ID $orderId");
+            }
+
+            $labelName = DpdLabel::DPD_LABEL_DIR . $order->getRef();
+            $labelName = $labelService->setLabelNameExtension($labelName);
+
+            $fileSystem = new Filesystem();
+
+            if (!$fileSystem->exists(DpdLabel::DPD_LABEL_DIR)) {
+                $fileSystem->mkdir(DpdLabel::DPD_LABEL_DIR);
+            }
+
+            $data = $this->validateForm($this->createForm(LabelGenerationForm::getName()))->getData();
+
             DpdLabel::setConfigValue('new_status', $data['new_status']);
+
+            $labelService->createLabel($order, $labelName, (float) $data['weight'], $retour, null, $data['new_status']);
         } catch (\Exception $e) {
-            return $this->generateRedirect(URL::getInstance()->absoluteUrl('/admin/order/update/' . $orderId, [
-                "err" => $e->getMessage(),
-                "tab" => 'bill'
-            ]));
+            $error = $e->getMessage();
         }
 
-        $err = $labelService->createLabel($order, $labelName, $data['weight'], $retour, null, $data['new_status']);
-
-        if (is_string($err)) {
-            return $this->generateRedirect(URL::getInstance()->absoluteUrl('/admin/order/update/' . $orderId, [
-                "err" => $err,
-                "tab" => 'bill'
-            ]));
+        if ($returnUrl) {
+            return new RedirectResponse($returnUrl);
         }
 
-        return $this->generateRedirect(URL::getInstance()->absoluteUrl('/admin/order/update/' . $orderId, [ "tab" => 'bill' ]));
+        return $this->generateRedirect(URL::getInstance()?->absoluteUrl('/admin/order/update/' . $orderId, [
+            "err" => $error,
+            "tab" => 'bill'
+        ]));
     }
 
     /**
      * @Route("/labels-file/{base64EncodedFilename}", name="_labels-file_download", methods="GET")
+     * @param $base64EncodedFilename
+     * @return Response
      */
-    public function downloadAction($base64EncodedFilename)
+    public function downloadAction(string $base64EncodedFilename, Translator $translator): Response
     {
         $fileName = base64_decode($base64EncodedFilename);
         [ 'filename' => $fileNameWithoutExt, 'dirname' => $dirname ] = pathinfo($fileName);
@@ -143,75 +141,94 @@ class LabelController extends BaseAdminController
 
         if (!empty($files) && file_exists($files[0])) {
             $fileName = $files[0];
-            header('Content-Description: File Transfer');
-            header('Content-Type: application/octet-stream');
-            header('Content-Disposition: attachment; filename="' . basename($fileName) . '"');
-            header('Expires: 0');
-            header('Cache-Control: must-revalidate');
-            header('Pragma: public');
-            header('Content-Length: ' . filesize($fileName));
-            readfile($fileName);
-        } else {
-            return $this->generateRedirect(URL::getInstance()->absoluteUrl("/admin/module/DpdLabel/labels"));
+
+            return new Response(
+                file_get_contents($fileName),
+                200,
+                [
+                    'Content-Description' => 'File Transfer',
+                    'Content-Type' => 'application/octet-stream',
+                    'Content-Disposition' => 'attachment; filename="' . basename($fileName) . '"',
+                    'Expires: 0',
+                    'Cache-Control' => 'must-revalidate',
+                    'Pragma' => 'public',
+                    'Content-Length' => filesize($fileName),
+            ]);
         }
 
-        return $this->generateRedirect(URL::getInstance()->absoluteUrl("/admin/module/DpdLabel/labels"));
+        return new Response(
+            $translator->trans(
+                "Label file %fileName was not found",
+                [ '%fileName' => $fileName ],
+                DpdLabel::DOMAIN_NAME
+            ),
+            Response::HTTP_NOT_FOUND
+        );
     }
 
     /**
      * @Route("/getLabel/{orderRef}", name="_get_label", methods="GET")
      */
-    public function getLabelAction($orderRef, Request $request)
+    public function getLabelAction($orderRef, Request $request, Translator $translator): Response
     {
-        $labelDir = DpdLabel::DPD_LABEL_DIR;
-        $file = $labelDir . DS . $orderRef;
-        $download = $request->get('download');
+        $file = DpdLabel::DPD_LABEL_DIR . $orderRef;
 
         $files = glob($file.'.*');
 
         if (!empty($files) && file_exists($files[0])) {
             $response = new BinaryFileResponse($files[0]);
 
-            if ($download) {
+            if ($request->get('download')) {
                 $response->setContentDisposition(
                     ResponseHeaderBag::DISPOSITION_ATTACHMENT,
-                    $orderRef . '.pdf'
+                    basename($files[0])
                 );
             }
 
             return $response;
         }
 
-        return '';
+        return new Response(
+            $translator->trans(
+                "Label not found for order ref. %ref",
+                [ '%ref' => $orderRef ],
+                DpdLabel::DOMAIN_NAME
+            ),
+            Response::HTTP_NOT_FOUND
+        );
     }
 
     /**
-     * @return mixed|\Symfony\Component\HttpFoundation\Response
+     * @param Request $request
+     * @return Response
      * @throws \Propel\Runtime\Exception\PropelException
      * @Route("/deleteLabel", name="_delete_label", methods="GET")
      */
-    public function deleteLabelAction()
+    public function deleteLabelAction(Request $request): Response
     {
         if (null !== $response = $this->checkAuth(AdminResources::ORDER, [], AccessManager::UPDATE)) {
             return $response;
         }
 
-        $orderId = $this->getRequest()->get("orderId");
+        $orderId = $request->get("orderId");
+        $returnUrl = $request->get("return_url");
 
-        $labelDir = DpdLabel::DPD_LABEL_DIR;
+        if (null !== $label = DpdlabelLabelsQuery::create()->filterByOrderId($orderId)->findOne()) {
+            $fs = new Filesystem();
 
-        $label = DpdlabelLabelsQuery::create()->filterByOrderId($orderId)->findOne();
+            $labelName = DpdLabel::DPD_LABEL_DIR . $label->getOrder()->getRef();
 
-        $fs = new Filesystem();
+            foreach (glob($labelName . '.*') as $filename) {
+                $fs->remove($filename);
+            }
 
-        $labelName = $labelDir . DS . $label->getOrder()->getRef();
-
-        foreach (glob($labelName.'.*') as $filename) {
-            $fs->remove($filename);
+            $label->delete();
         }
 
-        $label->delete();
+        if ($returnUrl) {
+            return new RedirectResponse($returnUrl);
+        }
 
-        return $this->generateRedirect(URL::getInstance()->absoluteUrl($this->getRequest()->get("redirect_url")));
+        return $this->generateRedirect(URL::getInstance()?->absoluteUrl($request->get("redirect_url")));
     }
 }
